+++ Entities/HrdCompanyInfo.cs
@@ -7,7 +7,7 @@ namespace POS_API.Entities
     {
         [Key]
         [Required]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public bool IsActive { get; set; } = true;
 
@@ -18,17 +18,17 @@ namespace POS_API.Entities
         public DateTime CreatedAt { get; set; }
 
 
-        public long? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
 
         public DateTime? UpdatedAt { get; set; }
 
 
-        public long? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
         public DateTime? DeletedAt { get; set; }
 
 
-        public long? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
         public bool? CompanyType { get; set; }
 
         public string? HeadOfficeId { get; set; }
