+++ Entities/Master/POSUnit.cs
@@ -6,7 +6,7 @@ namespace POS_API.Entities.Master
     {
         [Required]
         [Key]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
 
         public bool IsActive { get; set; } = true;
@@ -19,17 +19,17 @@ namespace POS_API.Entities.Master
         public DateTime CreatedAt { get; set; }
 
 
-        public long? CreatedBy { get; set; }
-        public long? CompanyId { get; set; }
+        public int? CreatedBy { get; set; }
+        public int? CompanyId { get; set; }
 
         public DateTime? UpdatedAt { get; set; }
 
 
-        public long? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
         public DateTime? DeletedAt { get; set; }
 
 
-        public long? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
         [Required]
         [MaxLength(50)]
         public string? Name { get; set; }
