+++ Entities/Purchase/POSPurchaseMaster.cs
@@ -8,23 +8,23 @@ namespace POS_API.Entities.Purchase
     {
         [Key]
         [Required]
-        public long Id { get; set; }
+        public int Id { get; set; }
         public bool IsActive { get; set; } = true;
         public bool IsDeleted { get; set; } = false;
 
         [Required]
         public DateTime CreatedAt { get; set; }
 
-        public long? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
         public DateTime? UpdatedAt { get; set; }
 
 
-        public long? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
         public DateTime? DeletedAt { get; set; }
 
 
-        public long? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
         [Required]
         [MaxLength(50)]
         public string? PurchaseNo { get; set; }
@@ -37,7 +37,7 @@ namespace POS_API.Entities.Purchase
 
 
         [Required]
-        public long SupplierId { get; set; }
+        public int SupplierId { get; set; }
